requirements.txt
django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-redis==6.0.0
pymysql==1.1.2
mysqlclient==2.2.8
python-dotenv==1.2.2
pillow==12.2.0
redis==7.4.0
minio==7.2.20
elasticsearch==8.17.2
gunicorn==23.0.0
paramiko==4.0.0
requests==2.33.1
